+++ src/staking/nominate.ts
@@ -1,27 +1,16 @@
 import Metadata from '@polkadot/metadata';
 import { createType, TypeRegistry } from '@polkadot/types';
-import { SignerPayloadJSON } from '@polkadot/types/types';
 
 import { EXTRINSIC_VERSION, ONE_SECOND } from '../util/constants';
 import { TxInfoNominate } from './stakingTxTypeUtils';
-
-/**
- * JSON format for an unsigned transaction
- */
-export interface UnsignedNominateTransaction extends SignerPayloadJSON {
-  /**
-   * The SCALE-encoded metadata, as a hex string. Can be retrieved via the RPC
-   * call `state_getMetadata`.
-   */
-  metadataRpc: string;
-}
+import { UnsignedTransaction } from '../util/interfaces';
 
 /**
  * Construct a transaction to nominate. This must be called by the _Controller_ account.
  *
  * @param info - Information required to construct the transaction.
  */
-export function nominate(info: TxInfoNominate): UnsignedNominateTransaction {
+export function nominate(info: TxInfoNominate): UnsignedTransaction {
   const registry = new TypeRegistry();
   const metadata = new Metadata(registry, info.metadataRpc);
